🚚  rename: formTypes 타입 정리 #44
