+++ src/common/types/formTypes.ts
@@ -1,0 +1,16 @@
+export interface AuthFormType {
+  email: string;
+  name: string;
+  phoneNumber: string;
+  password: string;
+  passwordConfirmation: string;
+}
+
+export type AccountDeletionFormType = Pick<AuthFormType, "password"> & { reason: string };
+export type LoginFormType = Pick<AuthFormType, "email" | "password"> & { reason: string };
+export type PasswordSettingFormType = Pick<AuthFormType, "password"> & {
+  originalPassword: string;
+  passwordConfirmation: string;
+};
+
+export type InputType = "text" | "email" | "password";
