+++ src/pages/Login/components/LoginForm.tsx
@@ -1,7 +1,7 @@
 import Button from "@/common/components/atoms/Button";
 import FormInputField from "@/common/components/molecules/FormInputField";
 import useLoginMutation from "@/pages/Login/api/useLoginMutation";
-import { LoginFormType } from "@/type";
+import { LoginFormType } from "@/common/types/formTypes";
 import { validationRules } from "@/validationRules";
 import { useForm } from "react-hook-form";
